refactor(BookForm): Made BookForm a stateless component to only render the form
