+++ src/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Header from './Header';
 import BookList from '../containers/BookList';
-import BookForm from '../containers/BookForm';
+import BookFormControl from '../containers/BookFormControl';
 import './App.scss';
 
 function App() {
@@ -9,7 +9,7 @@ function App() {
     <main className="App">
       <Header />
       <BookList />
-      <BookForm />
+      <BookFormControl />
     </main>
   );
 }
