+++ src/components/App.scss
@@ -66,18 +66,21 @@ body {
   height: 170px;
   border-radius: 4px;
   border: solid 1px #e8e8e8;
+  justify-content: space-between;
   box-sizing: border-box;
   background-color: #ffffff;
   display: flex;
+  flex-wrap: nowrap;
   margin: 8px 100px;
   padding: 32px;
 
   .book-info {
     display: flex;
     flex-direction: column;
+    flex-grow: 1;
+    width: 0%;
     justify-content: space-between;
     text-align: left;
-    flex-grow: 1;
 
     .category {
       font-size: 0.9rem;
@@ -89,6 +92,9 @@ body {
     h3 {
       font-size: 1.4rem;
       font-family: "Roboto Slab", serif;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+      overflow: hidden;
       margin: 4px 0 0;
       color: #121212;
     }
@@ -124,7 +130,8 @@ body {
   .display-progress {
     display: flex;
     flex-direction: row;
-    flex-grow: 0;
+    flex-grow: 1;
+    width: 0%;
     justify-content: center;
 
     svg {
@@ -158,14 +165,12 @@ body {
       margin-left: 24px;
 
       .number {
-        // font-family: "Roboto", serif;
         font-size: 2rem;
       }
 
       .completed {
         color: #121212;
         font-size: 0.9rem;
-
         opacity: 0.5;
       }
     }
@@ -174,6 +179,8 @@ body {
   .update-progress {
     display: flex;
     flex-direction: column;
+    flex-grow: 1;
+    width: 0%;
     text-align: left;
 
     .page-label {
@@ -195,6 +202,7 @@ body {
       border: 1px solid #3f90ff;
       border-radius: 3px;
       margin-top: auto;
+      // max-width: 190px;
       text-transform: uppercase;
       letter-spacing: 0.5px;
       color: #ffffff;
